Neutrino CMenuWidget: make page_start[] int, avoids lots of casts

+++ tuxbox/neutrino/src/gui/widget/menue.cpp
@@ -434,8 +434,8 @@ int CMenuWidget::exec(CMenuTarget* parent, const std::string &)
 
 						if (item->isSelectable())
 						{
-							if ((pos < (int)page_start[current_page + 1]) &&
-							    (pos >= (int)page_start[current_page]))
+							if ((pos < page_start[current_page + 1]) &&
+							    (pos >= page_start[current_page]))
 							{ // Item is currently on screen
 								//clear prev. selected
 								items[selected]->paint(false);
@@ -504,13 +504,13 @@ int CMenuWidget::exec(CMenuTarget* parent, const std::string &)
 					{
 						if (current_page)
 						{
-							pos = (int)page_start[current_page] - 1;
+							pos = page_start[current_page] - 1;
 							for (unsigned int count = pos; count > 0; count--)
 							{
 								CMenuItem* item = items[pos];
 								if (item->isSelectable())
 								{
-									if ((pos < (int)page_start[current_page + 1]) && (pos >= (int)page_start[current_page]))
+									if ((pos < page_start[current_page + 1]) && (pos >= page_start[current_page]))
 									{
 										items[selected]->paint(false);
 										item->paint(true);
@@ -534,7 +534,7 @@ int CMenuWidget::exec(CMenuTarget* parent, const std::string &)
 								CMenuItem* item = items[pos];
 								if (item->isSelectable())
 								{
-									if ((pos < (int)page_start[current_page + 1]) && (pos >= (int)page_start[current_page]))
+									if ((pos < page_start[current_page + 1]) && (pos >= page_start[current_page]))
 									{
 										items[selected]->paint(false);
 										item->paint(true);
@@ -553,7 +553,7 @@ int CMenuWidget::exec(CMenuTarget* parent, const std::string &)
 					}
 					else if (msg == g_settings.key_menu_pagedown)
 					{
-						pos = (int)page_start[current_page + 1];
+						pos = page_start[current_page + 1];
 						if (pos >= (int)items.size())
 							pos = items.size() - 1;
 						for (unsigned int count = pos; count < items.size(); count++)
@@ -561,7 +561,7 @@ int CMenuWidget::exec(CMenuTarget* parent, const std::string &)
 							CMenuItem* item = items[pos];
 							if (item->isSelectable())
 							{
-								if ((pos < (int)page_start[current_page + 1]) && (pos >= (int)page_start[current_page]))
+								if ((pos < page_start[current_page + 1]) && (pos >= page_start[current_page]))
 								{
 									items[selected]->paint(false);
 									item->paint(true);
@@ -703,9 +703,9 @@ void CMenuWidget::paintItems()
 	//Item not currently on screen
 	if (selected >= 0)
 	{
-		while(selected < (int)page_start[current_page])
+		while(selected < page_start[current_page])
 			current_page--;
-		while(selected >= (int)page_start[current_page + 1])
+		while(selected >= page_start[current_page + 1])
 			current_page++;
 	}
 
@@ -717,7 +717,7 @@ void CMenuWidget::paintItems()
 	}
 	frameBuffer->paintBoxRel(x,item_start_y, width,item_height, COL_MENUCONTENT_PLUS_0);
 	int ypos=item_start_y;
-	for (unsigned int count = 0; count < items.size(); count++)
+	for (int count = 0; count < (int)items.size(); count++)
 	{
 		CMenuItem* item = items[count];
